Swap order of DeflateStream and BufferedStream in our messagestream
DeflateStream was blocking in netstandard in the other order. No idea why.
I think they were only in the previous order because stock BufferedStream
doesn't have an option to not close the parent stream (which is the NetworkStream
in our new order). But we've forked BufferedStream to add the option for now.