Fixes a bug caused by exceptions in DataStreams.
This change fixes a bug where the remote would have 100% CPU load and
eventually stop servicing request when the sender throws an exception in
the `DataStream`.

The issue was the return value of the `read()` method was not
respected. The doc for this method says:
```
<returns>The number of bytes read into <paramref name="buffer" />. This might be less than the number of bytes requested if that many bytes are not available, or it might be zero if the end of the stream is reached.</returns>
```
The fix is to simply detect a return value of `0` indicating the stream
is closed rather than continue to try and read from the stream.

Fixes: https://github.com/OctopusDeploy/Halibut/issues/141

May be related to: https://github.com/OctopusDeploy/Halibut/issues/117