Run tests only for listening
Polling tentacles don't have "parallel" requests in the same way as
Listening tentacles, as there is a single queue from which the polling
tentacle will request work. This means that the polling test cases will
actually run the requests sequentially. Not only does this not test the
expected behaviour, but it also runs slower than expected on the build
server, causing the tests the consistently fail via timeout.

As running this test does not provide value on polling connection types,
only run the test on listening connections.